v1

+++ src/main/java/com/example/pensionatdb/controllers/kundController.java
@@ -2,21 +2,26 @@ package com.example.pensionatdb.controllers;
 
 import com.example.pensionatdb.models.Bokning;
 import com.example.pensionatdb.models.Kund;
+import com.example.pensionatdb.repos.bokningRepo;
 import com.example.pensionatdb.repos.kundRepo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
+
 @RestController
 public class kundController {
 
     private static final Logger log = LoggerFactory.getLogger(bokningController.class);
 
     private final kundRepo kr;
+    private final bokningRepo br;
 
-    public kundController(kundRepo kr) {
+    public kundController(kundRepo kr, bokningRepo br) {
         this.kr = kr;
+        this.br = br;
     }
 
     @RequestMapping("kund")
@@ -35,12 +40,23 @@ public class kundController {
         return kr.findAll();
     }
 
-    @RequestMapping("kund/{id}/delete")
-    public List<Kund> deleteById(@PathVariable Long id){
-        kr.deleteById(id);
-        log.info("kund deleted with id "+ id);
+
+
+    @RequestMapping("/kund/{id}/delete")
+    public List<Kund> deleteById(@PathVariable Long id) {
+        Optional<Kund> kundOptional = kr.findById(id);
+        if (kundOptional.isPresent()) {
+            Kund kund = kundOptional.get();
+            List<Bokning> bokningar = br.findByKund(kund);
+            if (bokningar.isEmpty()) {
+                kr.deleteById(id);
+                log.info("Kund deleted with id " + id);
+            }
+            else log.warn("Kund with id " + id + " cannot be deleted as they have a bokning .");
+        }
         return kr.findAll();
     }
+
 
     @RequestMapping("kundNameChange/{id}/{namn}")
     public List<Kund> updateName(@PathVariable Long id,@PathVariable String namn){
